Broke out table catcher module
moved table catcher module into new file tableCatcherModule.js.
